// apps/web-svelte/src/lib/api/generated/model/videoDirectPlayProfileRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface VideoDirectPlayProfileRequest {
  /** @nullable */
  type: string | null;
  /** @nullable */
  container: string | null;
  /** @nullable */
  videoCodec: string | null;
  /** @nullable */
  audioCodec: string | null;
}