// apps/web-svelte/src/lib/api/generated/model/videoPlaybackProfileRequest.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Prismedia.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { VideoDirectPlayProfileRequest } from './videoDirectPlayProfileRequest';

export interface VideoPlaybackProfileRequest {
  /**
     * @nullable
     * @pattern ^-?(?:0|[1-9]\d*)$
     */
  maxStreamingBitrate: number | string | null;
  /** @nullable */
  directPlayProfiles: VideoDirectPlayProfileRequest[] | null;
}